initial 1.5 upload

+++ xbmc/cores/dvdplayer/Codecs/ffmpeg/build_xbmc_win32.sh
@@ -1,23 +1,51 @@
-#!/bin/bash 
-rm -r .libs
-make distclean
+#!/bin/bash
 
-./configure \
---extra-cflags="-D_XBOX -fno-common" \
---disable-shared \
+if [ "$1" == "clean" ]
+then
+  if [ -d .libs ]
+  then
+    rm -r .libs
+  fi
+  make distclean
+fi
+
+if [ ! -d .libs ]; then
+  mkdir .libs
+fi
+
+OPTIONS="
+--enable-shared \
 --enable-memalign-hack \
 --enable-gpl \
 --enable-w32threads \
 --enable-postproc \
 --enable-zlib \
---enable-static \
+--disable-static \
 --disable-altivec \
 --disable-muxers \
 --disable-encoders \
---disable-ipv6 \
---disable-debug && 
+--disable-debug \
+--disable-ffplay \
+--disable-ffserver \
+--disable-ffmpeg \
+--disable-ffprobe \
+--disable-devices \
+--enable-muxer=spdif \
+--enable-muxer=adts \
+--enable-encoder=ac3 \
+--enable-encoder=aac \
+--enable-runtime-cpudetect \
+--disable-debug \
+--disable-doc"
+
+./configure --extra-cflags="-fno-common -Iinclude-xbmc-win32/dxva2" --extra-ldflags="-L/xbmc/system/players/dvdplayer" ${OPTIONS} &&
  
-make -j3 && 
-mkdir .libs &&
-cp lib*/*.a .libs/ &&
-mv .libs/swscale-0.dll .libs/swscale-0.6.1.dll
+make &&
+cp lib*/*.dll .libs/ &&
+mv .libs/swscale-0.dll .libs/swscale-0.6.1.dll &&
+cp .libs/avcodec-52.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/avcore-0.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/avformat-52.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/avutil-50.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/postproc-51.dll /xbmc/system/players/dvdplayer/ &&
+cp .libs/swscale-0.6.1.dll /xbmc/system/players/dvdplayer/
